Switch from injecting phdrs to relying linker
Assume the ehdr and phdrs will be in a PT_LOAD section, and that
__ehdr_start will point to the ehdr. This is the default behavior of
LLD.

We incur this dependency on linker behavior (or the need for a linker
script) in order to remove the additional build step of injecting phdrs.

We keep the phdr injection code around for cases where it is still
useful.